Success: passed test optional-1

+++ 07 - belajar-fundamental-backend-dengan-javascript/open-music-03/src/modules/collaborations/application/use-case/delete-collaboration.use-case.ts
@@ -1,12 +1,12 @@
-import { NotFoundError } from '../../../../shared/errors/app-error';
+import { ForbiddenError, NotFoundError } from '../../../../shared/errors/app-error';
 import { CollaborationRepository } from '../../infrasctructure/collaboration.repository';
 import { DeleteCollaborationDto } from '../dto/delete-collaboration.dto';
 
 export class DeleteCollaborationUseCase {
   constructor(private readonly collaborationRepository: CollaborationRepository) {}
 
-  async execute(payload: DeleteCollaborationDto) {
-    const { userId, playlistId } = payload;
+  async execute(payload: DeleteCollaborationDto & { authId: string }) {
+    const { authId, userId, playlistId } = payload;
     const collaboration = await this.collaborationRepository.findByPlaylistIdAndUserId({
       playlistId,
       userId,
@@ -14,6 +14,10 @@ export class DeleteCollaborationUseCase {
 
     if (!collaboration) {
       throw new NotFoundError('Collaboration is not found');
+    }
+
+    if (authId === userId) {
+      throw new ForbiddenError('Forbidden access');
     }
 
     const deleted = await this.collaborationRepository.delete(collaboration);
